pyproject.toml
[build-system]
requires = ["hatchling>=1.26", "hatch-vcs"]
build-backend = "hatchling.build"


[project]
name = "discuss-nutshell"
authors = [
  { name = "Carol Willing", email = "willingc@gmail.com" },
]
description = "Query Discourse and summarize threads"
readme = "README.md"
license = "MIT"
license-files = ["LICENSE"]
requires-python = ">=3.12"
classifiers = [
  "Development Status :: 1 - Planning",
  "Intended Audience :: Science/Research",
  "Intended Audience :: Developers",
  "Operating System :: OS Independent",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python :: 3.14",
  "Topic :: Scientific/Engineering",
  "Typing :: Typed",
]
dynamic = ["version"]
dependencies = [
    "sqlite_utils",
    "requests",
    "pandas>=2.3.3",
    "beautifulsoup4>=4.14.2",
    "gradio>=5.50.0",
    "datasette>=0.65.2",
    "modal>=1.2.4",
    "jupyter>=1.1.1",
    "python-dotenv>=1.2.1",
    "pyarrow>=22.0.0",
    "seaborn>=0.13.2",
    "matplotlib>=3.10.7",
    "openai>=2.8.1",
    "anthropic>=0.74.1",
    "google-genai>=1.52.0",
]

[project.urls]
Homepage = "https://github.com/willingc/discuss-nutshell"
"Bug Tracker" = "https://github.com/willingc/discuss-nutshell/issues"
Discussions = "https://github.com/willingc/discuss-nutsshell/discussions"
Changelog = "https://github.com/willingc/discuss-nutshell/releases"


[dependency-groups]
test = [
  "pytest >=6",
  "pytest-cov >=3",
  "nox>=2025.11.12",
]
dev = [
  { include-group = "test" },
  "jupyter",
  "ipykernel",
  "requests",
]
docs = [
    "markdown>=3.9",
    "mdx-include>=1.4.2",
    "mkdocs-material>=9.1.19",
    "mkdocs>=1.1.2",
    "mkdocstrings-python>=1.18.2",
    "pyyaml>=6.0.1",
]
nb = [
  "jupyter",
  "ipykernel",
  "pandas",
  "nbstripout>=0.8.2",
]
[tool.hatch]
version.source = "vcs"
build.hooks.vcs.version-file = "src/discuss_nutshell/_version.py"

[tool.hatch.envs.default]
# duplicate for now since hatch doesn't support groups yet
dependencies = []
scripts.test = "pytest {args}"


[tool.pytest.ini_options]
minversion = "6.0"
addopts = ["-ra", "--showlocals", "--strict-markers", "--strict-config"]
xfail_strict = true
filterwarnings = [
  "error",
]
log_level = "INFO"
testpaths = [
  "tests",
]


[tool.coverage]
run.source = ["discuss_nutshell"]
report.exclude_also = [
  '\.\.\.',
  'if typing.TYPE_CHECKING:',
]

[tool.mypy]
files = ["src", "tests"]
python_version = "3.12"
warn_unused_configs = true
strict = true
enable_error_code = ["ignore-without-code", "redundant-expr", "truthy-bool"]
warn_unreachable = true
disallow_untyped_defs = false
disallow_incomplete_defs = false

[[tool.mypy.overrides]]
module = "discuss_nutshell.*"
disallow_untyped_defs = true
disallow_incomplete_defs = true

[[tool.mypy.overrides]]
module = ["gradio.*", "google.*"]
ignore_missing_imports = true


[tool.ruff]

[tool.ruff.lint]
extend-select = [
  "ARG",    # flake8-unused-arguments
  "B",      # flake8-bugbear
  "BLE",    # flake8-blind-except
  "C4",     # flake8-comprehensions
  "DTZ",    # flake8-datetimez
  "EM",     # flake8-errmsg
  "EXE",    # flake8-executable
  "FA",     # flake8-future-annotations
  "FLY",    # flynt
  "FURB",   # refurb
  "G",      # flake8-logging-format
  "I",      # isort
  "ICN",    # flake8-import-conventions
  "ISC",    # flake8-implicit-str-concat
  "LOG",    # flake8-logging
  "NPY",    # NumPy specific rules
  "PD",     # pandas-vet
  "PERF",   # perflint
  "PGH",    # pygrep-hooks
  "PIE",    # flake8-pie
  "PL",     # pylint
  "PT",     # flake8-pytest-style
  "PTH",    # flake8-use-pathlib
  "PYI",    # flake8-pyi
  "Q",      # flake8-quotes
  "RET",    # flake8-return
  "RSE",    # flake8-raise
  "RUF",    # Ruff-specific
  "SIM",    # flake8-simplify
  "SLOT",   # flake8-slots
  "T10",    # flake8-debugger
  "T20",    # flake8-print
  "TC",     # flake8-type-checking
  "TRY",    # tryceratops
  "UP",     # pyupgrade
  "YTT",    # flake8-2020
]
ignore = [
  "PLR09",    # Too many <...>
  "PLR2004",  # Magic value used in comparison
]
# Uncomment if using a _compat.typing backport
# typing-modules = ["discuss_nutshell._compat.typing"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["T20"]
"noxfile.py" = ["T20"]
"docs/examples/**" = ["T20"]


[tool.pylint]
py-version = "3.12"
ignore-paths = [".*/_version.py"]
reports.output-format = "colorized"
similarities.ignore-imports = "yes"
messages_control.disable = [
  "design",
  "fixme",
  "line-too-long",
  "missing-module-docstring",
  "missing-function-docstring",
  "wrong-import-position",
  "no-member",
]
